+++ src/main/java/ladder/domain/Players.java
@@ -26,12 +26,13 @@ public class Players {
     public List<String> getPlayerNames() {
         return players.stream()
                 .map(Player::getName)
+                .map(Name::getName)
                 .collect(Collectors.toList());
     }
 
     public int getNameMaxLength() {
         return this.players.stream()
-                .map(player -> player.getName().length())
+                .map(player -> player.getName().getName().length())
                 .max(Integer::compareTo)
                 .orElseThrow(() -> new IllegalStateException("플레이어가 존재하지 않습니다."));
     }
